+++ packages/examples/resources/appPlayground/hello.ts
@@ -1,6 +1,5 @@
-import { sayFoo } from './tester.js';
-
 function sayHello(): string {
+    // intentionally erroneous to test import resolution
     console.log(sayFoo());
     return 'Hello';
 };