+++ packages/examples/src/appPlayground/config.ts
@@ -3,10 +3,13 @@
  * Licensed under the MIT License. See LICENSE in the package root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import { Uri } from 'vscode';
+import * as vscode from 'vscode';
+import { LogLevel } from 'vscode/services';
+import { RegisteredFileSystemProvider, registerFileSystemOverlay, RegisteredMemoryFile } from '@codingame/monaco-vscode-files-service-override';
 import getConfigurationServiceOverride from '@codingame/monaco-vscode-configuration-service-override';
 import getKeybindingsServiceOverride from '@codingame/monaco-vscode-keybindings-service-override';
 import getLifecycleServiceOverride from '@codingame/monaco-vscode-lifecycle-service-override';
+import getLocalizationServiceOverride from '@codingame/monaco-vscode-localization-service-override';
 import getBannerServiceOverride from '@codingame/monaco-vscode-view-banner-service-override';
 import getStatusBarServiceOverride from '@codingame/monaco-vscode-view-status-bar-service-override';
 import getTitleBarServiceOverride from '@codingame/monaco-vscode-view-title-bar-service-override';
@@ -14,61 +17,55 @@ import getExplorerServiceOverride from '@codingame/monaco-vscode-explorer-servic
 import getRemoteAgentServiceOverride from '@codingame/monaco-vscode-remote-agent-service-override';
 import getEnvironmentServiceOverride from '@codingame/monaco-vscode-environment-service-override';
 import getSecretStorageServiceOverride from '@codingame/monaco-vscode-secret-storage-service-override';
-import { LogLevel } from 'vscode/services';
+import getStorageServiceOverride from '@codingame/monaco-vscode-storage-service-override';
+import getSearchServiceOverride from '@codingame/monaco-vscode-search-service-override';
+
+// this is required syntax highlighting
+import '@codingame/monaco-vscode-typescript-basics-default-extension';
+import '@codingame/monaco-vscode-typescript-language-features-default-extension';
+import '@codingame/monaco-vscode-search-result-default-extension';
+
+import '../../resources/vsix/open-collaboration-tools.vsix';
+
+import { createDefaultLocaleConfiguration } from 'monaco-languageclient/vscode/services';
+import { configureMonacoWorkers, createDefaultWorkspaceFile } from '../common/client/utils.js';
+import helloTsCode from '../../resources/appPlayground/hello.ts?raw';
+import testerTsCode from '../../resources/appPlayground/tester.ts?raw';
 import { WrapperConfig } from 'monaco-editor-wrapper';
-import { configureMonacoWorkers } from '../../common/client/utils.js';
-import { ClangdWorkerHandler } from './workerHandler.js';
 import { defaultHtmlAugmentationInstructions, defaultViewsInit } from 'monaco-editor-wrapper/vscode/services';
 
-export const createWrapperConfig = async (config: {
-    htmlContainer: HTMLElement,
-    workspaceUri: Uri,
-    workspaceFileUri: Uri,
-    clangdWorkerHandler: ClangdWorkerHandler,
-    lsMessageLocalPort: MessagePort
-}): Promise<WrapperConfig> => {
-    return {
+export type ConfigResult = {
+    wrapperConfig: WrapperConfig
+    workspaceFile: vscode.Uri;
+    helloTsUri: vscode.Uri;
+    testerTsUri: vscode.Uri;
+};
+
+export const configure = (htmlContainer?: HTMLElement): ConfigResult => {
+    const workspaceFile = vscode.Uri.file('/workspace/.vscode/workspace.code-workspace');
+
+    const wrapperConfig: WrapperConfig = {
         $type: 'extended',
-        htmlContainer: config.htmlContainer,
+        id: 'AAP',
         logLevel: LogLevel.Debug,
-        languageClientConfigs: {
-            LANGUAGE_ID: {
-                name: 'Clangd WASM Language Server',
-                connection: {
-                    options: {
-                        $type: 'WorkerDirect',
-                        worker: await config.clangdWorkerHandler.createWorker(),
-                        messagePort: config.lsMessageLocalPort
-                    }
-                },
-                restartOptions: {
-                    retries: 5,
-                    timeout: 1000,
-                    keepWorker: true
-                },
-                clientOptions: {
-                    documentSelector: ['cpp'],
-                    workspaceFolder: {
-                        index: 0,
-                        name: 'workspace',
-                        uri: config.workspaceUri
-                    }
-                }
-            }
-        },
+        htmlContainer,
         vscodeApiConfig: {
             serviceOverrides: {
                 ...getConfigurationServiceOverride(),
                 ...getKeybindingsServiceOverride(),
                 ...getLifecycleServiceOverride(),
+                ...getLocalizationServiceOverride(createDefaultLocaleConfiguration()),
                 ...getBannerServiceOverride(),
                 ...getStatusBarServiceOverride(),
                 ...getTitleBarServiceOverride(),
                 ...getExplorerServiceOverride(),
                 ...getRemoteAgentServiceOverride(),
                 ...getEnvironmentServiceOverride(),
-                ...getSecretStorageServiceOverride()
+                ...getSecretStorageServiceOverride(),
+                ...getStorageServiceOverride(),
+                ...getSearchServiceOverride()
             },
+            enableExtHostWorker: true,
             viewsConfig: {
                 viewServiceType: 'ViewsService',
                 htmlAugmentationInstructions: defaultHtmlAugmentationInstructions,
@@ -77,7 +74,7 @@ export const createWrapperConfig = async (config: {
             workspaceConfig: {
                 enableWorkspaceTrust: true,
                 windowIndicator: {
-                    label: 'mlc-clangd-example',
+                    label: 'mlc-app-playground',
                     tooltip: '',
                     command: ''
                 },
@@ -88,31 +85,32 @@ export const createWrapperConfig = async (config: {
                         return true;
                     },
                     workspace: {
-                        workspaceUri: config.workspaceFileUri
-                    },
+                        workspaceUri: workspaceFile
+                    }
                 },
                 configurationDefaults: {
-                    'window.title': 'mlc-clangd-exampled${separator}${dirty}${activeEditorShort}'
+                    'window.title': 'mlc-app-playground${separator}${dirty}${activeEditorShort}'
                 },
                 productConfiguration: {
-                    nameShort: 'mlc-clangd-example',
-                    nameLong: 'mlc-clangd-example'
+                    nameShort: 'mlc-app-playground',
+                    nameLong: 'mlc-app-playground'
                 }
             },
             userConfiguration: {
                 json: JSON.stringify({
                     'workbench.colorTheme': 'Default Dark Modern',
                     'editor.wordBasedSuggestions': 'off',
+                    'typescript.tsserver.web.projectWideIntellisense.enabled': true,
+                    'typescript.tsserver.web.projectWideIntellisense.suppressSemanticErrors': false,
                     'editor.guides.bracketPairsHorizontal': true,
-                    'editor.inlayHints.enabled': 'offUnlessPressed',
-                    'editor.quickSuggestionsDelay': 200,
+                    'oct.serverUrl': 'https://api.open-collab.tools/',
                     'editor.experimental.asyncTokenization': false
                 })
-            }
+            },
         },
         extensions: [{
             config: {
-                name: 'mlc-clangd-example',
+                name: 'mlc-app-playground',
                 publisher: 'TypeFox',
                 version: '1.0.0',
                 engines: {
@@ -123,5 +121,20 @@ export const createWrapperConfig = async (config: {
         editorAppConfig: {
             monacoWorkerFactory: configureMonacoWorkers
         }
+    };
+
+    const helloTsUri = vscode.Uri.file('/workspace/hello.ts');
+    const testerTsUri = vscode.Uri.file('/workspace/tester.ts');
+    const fileSystemProvider = new RegisteredFileSystemProvider(false);
+    fileSystemProvider.registerFile(new RegisteredMemoryFile(helloTsUri, helloTsCode));
+    fileSystemProvider.registerFile(new RegisteredMemoryFile(testerTsUri, testerTsCode));
+    fileSystemProvider.registerFile(createDefaultWorkspaceFile(workspaceFile, '/workspace'));
+    registerFileSystemOverlay(1, fileSystemProvider);
+
+    return {
+        wrapperConfig,
+        workspaceFile,
+        helloTsUri,
+        testerTsUri
     };
 };