+++ packages/examples/src/appPlayground/reactMain.tsx
@@ -3,38 +3,31 @@
  * Licensed under the MIT License. See LICENSE in the package root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import * as vscode from 'vscode';
-import { RegisteredFileSystemProvider, registerFileSystemOverlay, RegisteredMemoryFile } from '@codingame/monaco-vscode-files-service-override';
 import React, { StrictMode } from 'react';
 import ReactDOM from 'react-dom/client';
-import { MonacoEditorReactComp } from '@typefox/monaco-editor-react';
 import { MonacoEditorLanguageClientWrapper, TextChanges } from 'monaco-editor-wrapper';
-import { createWrapperConfig } from './config.js';
-import badPyCode from '../../../resources/python/bad.py?raw';
+import { MonacoEditorReactComp } from '@typefox/monaco-editor-react';
+import { configure } from './config.js';
+import { configurePostStart } from './common.js';
 
-export const runPythonReact = async () => {
-    const badPyUri = vscode.Uri.file('/workspace/bad.py');
-    const fileSystemProvider = new RegisteredFileSystemProvider(false);
-    fileSystemProvider.registerFile(new RegisteredMemoryFile(badPyUri, badPyCode));
-    registerFileSystemOverlay(1, fileSystemProvider);
+export const runApplicationPlaygroundReact = async () => {
 
     const onTextChanged = (textChanges: TextChanges) => {
         console.log(`Dirty? ${textChanges.isDirty}\ntext: ${textChanges.modified}\ntextOriginal: ${textChanges.original}`);
     };
-    const wrapperConfig = createWrapperConfig('/workspace', badPyCode, '/workspace/bad.py');
+    const configResult = configure();
     const root = ReactDOM.createRoot(document.getElementById('react-root')!);
 
     try {
         document.querySelector('#button-start')?.addEventListener('click', async () => {
             const App = () => {
                 return (
-                    <div style={{ 'height': '80vh', padding: '5px' }} >
+                    <div style={{ 'backgroundColor': '#1f1f1f' }}>
                         <MonacoEditorReactComp
-                            wrapperConfig={wrapperConfig}
-                            style={{ 'height': '100%' }}
+                            wrapperConfig={configResult.wrapperConfig}
                             onTextChanged={onTextChanged}
-                            onLoad={(wrapper: MonacoEditorLanguageClientWrapper) => {
-                                console.log(`Loaded ${wrapper.reportStatus().join('\n').toString()}`);
+                            onLoad={async (wrapper: MonacoEditorLanguageClientWrapper) => {
+                                await configurePostStart(wrapper, configResult);
                             }}
                             onError={(e) => {
                                 console.error(e);
@@ -43,16 +36,13 @@ export const runPythonReact = async () => {
                 );
             };
 
-            const strictMode = (document.getElementById('checkbox-strictmode')! as HTMLInputElement).checked;
-
-            if (strictMode) {
+            const checkElem = document.getElementById('checkbox-strictmode') as HTMLInputElement | null;
+            const strictMode = checkElem?.checked;
+            if (strictMode === true) {
                 root.render(<StrictMode><App /></StrictMode>);
             } else {
                 root.render(<App />);
             }
-        });
-        document.querySelector('#button-dispose')?.addEventListener('click', () => {
-            root.render([]);
         });
     } catch (e) {
         console.error(e);